Ajuste file manager
Se ha actualizado la version de Dile manager para el manejo de archivos por cliente.

+++ src/Http/PaginaController.php
@@ -32,18 +32,27 @@ protected $tenantName = null;
             $fqdn = $hostname->fqdn;
             $this->tenantName = explode(".", $fqdn)[0];
         }
- }
- dd($this->tenantName);
- }
 
+ }
 
  public function index(){
+  if(!$this->tenantName){
   $paginas = Page::all();
   $menu = Page::whereNull('page_id')->get();
-  $casta = DB::table('pages')->count();
+  $casta = Page::count();
   $user = Page::where('posti','=','1')->count();
+  }
+  else{
+  $paginas = \DigitalsiteSaaS\Pagina\Tenant\Page::all();
+  $menu = \DigitalsiteSaaS\Pagina\Tenant\Page::whereNull('page_id')->get();
+  $casta = \DigitalsiteSaaS\Pagina\Tenant\Page::count();
+  $user = \DigitalsiteSaaS\Pagina\Tenant\Page::where('posti','=','1')->count();
+  }
   return view('pagina::paginas.paginas')->with('paginas', $paginas)->with('user', $user)->with('casta', $casta)->with('menu', $menu);
- }
+ 
+}
+
+
 
  public function show(){
   $number = Auth::user()->id;
@@ -51,7 +60,25 @@ protected $tenantName = null;
   return View('pagina::paginas.crear-pagina')->with('user', $user)->with('number', $number);
  }
 
+ public function registrosaas(){
+  $tenantName = $this->tenantName;
+  return View('auth.register')->with('tenantName', $tenantName);
+ }
+
+  public function editardiagrama($id){
+  if(!$this->tenantName){
+  $diagramas = Diagrama::where('id', "=", $id)->get();
+  }else{
+  $diagramas = \DigitalsiteSaaS\Pagina\Tenant\Diagrama::where('id', "=", $id)->get();
+  }
+  return View('pagina::actualizar-diagrama')->with('diagramas', $diagramas);
+ }
+
+
+
  public function crearpagina(){
+
+  if(!$this->tenantName){
   $pagina = new Page;
   $pagina->page = Input::get('pagina');
   $pagina->slug = Str::slug($pagina->page);
@@ -92,12 +119,60 @@ protected $tenantName = null;
   $diagrama->bloqueficha1 = 'container';
   $diagrama->bloqueficha2 = 'container';
   $diagrama->save();
+ }
+ else{
+  $pagina = new \DigitalsiteSaaS\Pagina\Tenant\Page;
+  $pagina->page = Input::get('pagina');
+  $pagina->slug = Str::slug($pagina->page);
+  $pagina->description = Input::get('descripcion');
+  $pagina->visualizafoot = Input::get('visualizafoot');
+  $pagina->titulo = Input::get('titulo');
+  $pagina->palabras = Input::get('palabras');
+  $pagina->posti = Input::get('posti');
+  $pagina->nivel = Input::get('nivel');
+  $pagina->categoria = Input::get('categoria');
+  $pagina->sitio = Input::get('sitio');
+  $pagina->page_id = Input::get('DNI');
+  $pagina->save();
+  $diagrama = new \DigitalsiteSaaS\Pagina\Tenant\Diagrama;
+  $diagrama->posicionSD01 = 'row';
+  $diagrama->posicionSD1 = 'container';
+  $diagrama->posicionSD2 = 'row';
+  $diagrama->posicionSD02 = 'container';
+  $diagrama->posicionSD3 = 'container';
+  $diagrama->bloque = 'container';
+  $diagrama->bloqueSD1 = 'container';
+  $diagrama->bloqueSD2 = 'container';
+  $diagrama->bloqueSD3 = 'container';
+  $diagrama->bloqueSD4 = 'container';
+  $diagrama->bloqueSD5 = 'container';
+  $diagrama->bloqueSD6 = 'container';
+  $diagrama->bloqueSD7 = 'container';
+  $diagrama->bloqueSD8 = 'container';
+  $diagrama->bloqueSD9 = 'container';
+  $diagrama->posicionSD4 = 'col-xs-12 col-sm-12 col-md-3 col-lg-3';
+  $diagrama->posicionSD5 = 'col-xs-12 col-sm-12 col-md-6 col-lg-6';
+  $diagrama->posicionSD6 = 'col-xs-12 col-sm-12 col-md-3 col-lg-3';
+  $diagrama->posicionSD7 = 'container';
+  $diagrama->posicionSD8 = 'container';
+  $diagrama->posicionSD9 = 'row';
+  $diagrama->bloqueblog = 'container';
+  $diagrama->posicionSD03 = 'container';
+  $diagrama->bloqueficha1 = 'container';
+  $diagrama->bloqueficha2 = 'container';
+  $diagrama->save();
+  }
   return Redirect('gestion/paginas')->with('status', 'ok_create');
  }
 
  public function actualizar($id){
   $input = Input::all();
+  if(!$this->tenantName){
   $pagina = Page::find($id);
+  }
+  else{
+  $pagina = \DigitalsiteSaaS\Pagina\Tenant\Page::find($id);
+  }
   $pagina->page = Input::get('pagina');
   $pagina->slug = Str::slug($pagina->page);
   $pagina->visualizafoot = Input::get('visualizafoot');
@@ -112,8 +187,14 @@ protected $tenantName = null;
  }
 
  public function editar($id){
+  if(!$this->tenantName){
   $number = Auth::user()->id;
   $paginas = Page::find($id);
+}
+else{
+  $number = Auth::user()->id;
+  $paginas = \DigitalsiteSaaS\Pagina\Tenant\Page::find($id);
+}
   return view('pagina::paginas.editar-pagina')->with('paginas', $paginas)->with('number', $number);
  }
 
@@ -123,9 +204,9 @@ protected $tenantName = null;
  }
 
  public function eliminar($id){
-
+ if(!$this->tenantName){
   $res = Diagrama::where('id',$id)->delete();
-  $conteo = DB::table('pages')->where('page_id','=',$id)->count();
+  $conteo = Page::where('page_id','=',$id)->count();
   if($conteo == 0){
   $pagina = Page::find($id);
   $pagina->delete();
@@ -133,6 +214,18 @@ protected $tenantName = null;
   }
   else{
     return Redirect('/gestion/paginas')->with('status', 'ok_nodelete');
+  }
+ }else{
+  $res = \DigitalsiteSaaS\Pagina\Tenant\Diagrama::where('id',$id)->delete();
+  $conteo = \DigitalsiteSaaS\Pagina\Tenant\Page::where('page_id','=',$id)->count();
+  if($conteo == 0){
+  $pagina = \DigitalsiteSaaS\Pagina\Tenant\Page::find($id);
+  $pagina->delete();
+    return Redirect('/gestion/paginas')->with('status', 'ok_delete');
+  }
+  else{
+    return Redirect('/gestion/paginas')->with('status', 'ok_nodelete');
+  }
   }
  }
